Moved clerk buttons to dashboard
